refactor(channel)!: merge channel crates into one crate
BREAKING CHANGE: merge channel crates into one `xs-channel` crate
