+++ crates/channel-primitives/Cargo.toml
@@ -14,12 +14,11 @@ scale-info = { version = "2.1.1", default-features = false, features = [ "derive
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.27", default-features = false, version = "6.0.0" }
 sp-std     = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.27", default-features = false, version = "4.0.0" }
 
+xp-format = { path = "../format", default-features = false, version = "0.3.2" }
+
 # Optional frame specific deps
 frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.27", default-features = false, optional = true, version = "4.0.0-dev" }
-
-xcm = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.27", default-features = false, optional = true, version = "0" }
-
-xp-format = { path = "../format", default-features = false, version = "0.3.2" }
+xcm           = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.27", default-features = false, optional = true, version = "0" }
 
 [dev-dependencies]
 lazy_static = "1.4.0"
