+++ crates/channel/src/lib.rs
@@ -1,0 +1,5 @@
+#![feature(box_syntax)]
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod receiver;
+pub mod sender;
