+++ crates/channel/src/receiver/frame/sync.rs
@@ -1,7 +1,10 @@
-use super::{
-    handler_to_xbi_result, instruction_error_to_xbi_result, invert_destination_from_message,
+use crate::{
+    receiver::frame::{
+        handler_to_dispatch_info, handler_to_xbi_result, instruction_error_to_xbi_result,
+        invert_destination_from_message,
+    },
+    receiver::Receiver as ReceiverExt,
 };
-use crate::{frame::handler_to_dispatch_info, Receiver as ReceiverExt};
 use codec::Encode;
 use frame_support::pallet_prelude::DispatchResultWithPostInfo;
 use frame_system::{ensure_signed, Config};
@@ -14,16 +17,19 @@ use xp_channel::{
 };
 use xp_format::{XbiFormat, XbiMetadata, XbiResult};
 
-/// This is an asynchronous queue backed frame receiver, which expects some queue handler to transport the messages back via the transport layer,
-/// detaching the message handling part with the transport of the message.
-pub struct Receiver<T, Emitter, Queue, InstructionHandler> {
-    phantom: PhantomData<(T, Emitter, Queue, InstructionHandler)>,
+/// This is a synchronous backed Frame receiver
+/// It services the `REQ-REP` side of an async channel, that is to say, it receives a message, handles it, then responds with the result
+/// all in one step.
+pub struct Receiver<T, Sender, Emitter, Queue, InstructionHandler> {
+    phantom: PhantomData<(T, Sender, Emitter, Queue, InstructionHandler)>,
 }
 
-impl<T, Emitter, Queue, InstructionHandler> ReceiverExt
-    for Receiver<T, Emitter, Queue, InstructionHandler>
+#[cfg(feature = "frame")]
+impl<T, Sender, Emitter, Queue, InstructionHandler> ReceiverExt
+    for Receiver<T, Sender, Emitter, Queue, InstructionHandler>
 where
     T: Config,
+    Sender: crate::sender::Sender<Message>,
     Emitter: ChannelProgressionEmitter,
     Queue: Queueable<(Message, QueueSignal)>,
     InstructionHandler: XbiInstructionHandler<T::Origin>,
@@ -31,12 +37,11 @@ where
     type Origin = T::Origin;
     type Outcome = DispatchResultWithPostInfo;
 
-    /// Request should always run the instruction, and produce some info containing meters for the execution
+    /// Request should always run the instruction, and product some dispatch info containing meters for the execution
     fn handle_request(origin: &Self::Origin, msg: &mut XbiFormat) -> DispatchResultWithPostInfo {
         let _who = ensure_signed(origin.clone())?;
         let current_block: u32 = <frame_system::Pallet<T>>::block_number().unique_saturated_into();
 
-        // progress to delivered
         msg.metadata.timesheet.progress(current_block);
 
         Emitter::emit_received(Either::Left(msg));
@@ -45,47 +50,39 @@ where
 
         let xbi_id = msg.metadata.id::<T::Hashing>();
 
-        let instruction_result = InstructionHandler::handle(origin, msg);
+        let instruction_handle = InstructionHandler::handle(origin, msg);
 
-        let xbi_result = match &instruction_result {
+        let xbi_result = match &instruction_handle {
             Ok(info) => handler_to_xbi_result::<Emitter>(&xbi_id.encode(), info, msg),
             Err(e) => instruction_error_to_xbi_result(&xbi_id.encode(), e),
         };
 
-        // progress to executed
         msg.metadata.timesheet.progress(current_block);
 
         Emitter::emit_request_handled(
             &xbi_result,
             &msg.metadata,
-            &match &instruction_result {
+            &match &instruction_handle {
                 Ok(info) => Some(info.weight),
                 Err(e) => e.post_info.actual_weight,
             }
             .unwrap_or_default(),
         );
 
-        Queue::push((
-            Message::Response(xbi_result, msg.metadata.clone()),
-            QueueSignal::ResponseReceived,
-        ));
+        Sender::send(Message::Response(xbi_result, msg.metadata.clone()));
 
-        handler_to_dispatch_info(instruction_result)
+        handler_to_dispatch_info(instruction_handle)
     }
 
-    /// Response should delegate to the queue handler who would know about how to handle the message
+    // TODO: this should not have a queue anymore, we should provide some storage interface to write the result and add the cost.
+    /// Response should update the state of the storage checkout queues and notify the sender of completion
     fn handle_response(
         origin: &Self::Origin,
         msg: &XbiResult,
-        metadata: &XbiMetadata,
+        _metadata: &XbiMetadata,
     ) -> DispatchResultWithPostInfo {
         let _who = ensure_signed(origin.clone())?;
         Emitter::emit_received(Either::Right(msg));
-
-        Queue::push((
-            Message::Response(msg.clone(), metadata.clone()),
-            QueueSignal::ResponseReceived,
-        ));
 
         // TODO: add the cost of handling this response here
         Ok(Default::default())
