--- crates/sender/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-authors     = [ "t3rn <team@t3rn.io>" ]
-description = "A sender channel part of XBI"
-edition     = "2021"
-license     = "Apache-2.0"
-name        = "xs-sender"
-repository  = "https://github.com/t3rn/xbi"
-version     = "0.3.2"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-log = { version = "0.4", default-features = false }
-
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.27", default-features = false, version = "6.0.0" }
-
-xp-channel = { path = "../../crates/channel-primitives", default-features = false, version = "0.3" }
-xp-xcm     = { path = "../xcm-primitives", default-features = false, features = [ "frame" ], optional = true, version = "0.3" }
-
-# Optional frame specific deps
-frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.27", version = "4.0.0-dev", default-features = false, optional = true }
-frame-system  = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.27", version = "4.0.0-dev", default-features = false, optional = true }
-
-xcm = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.27", version = "0", default-features = false, optional = true }
-
-[dev-dependencies]
-lazy_static = "1.4.0"
-serial_test = "1.0.0"
-
-[features]
-default   = [ "std", "frame-std" ]
-frame     = [ "frame-support", "frame-system", "xp-xcm", "xcm" ]
-frame-std = [ "frame-support/std", "frame-system/std", "xp-xcm/std", "xp-xcm/frame-std", "xcm/std" ]
-std       = [ "sp-runtime/std", "xp-channel/std", "log/std" ]
